+++ app/src/main/java/com/silverknife/meizhi/MyApplication.kt
@@ -1,20 +1,19 @@
 package com.silverknife.meizhi
 
-import android.app.Application
-import android.view.Gravity
-
 import com.silvericekey.skutilslibrary.SKUtilsLibrary
+import com.silvericekey.skutilslibrary.base.BaseApplication
 import com.silvericekey.skutilslibrary.utils.HttpUtil
 
 /**
  * Created by silverknife on 2017/12/13.
  */
 
-class MyApplication : Application() {
+class MyApplication : BaseApplication() {
 
     override fun onCreate() {
         super.onCreate()
         SKUtilsLibrary.init(this)
         HttpUtil.init("http://football.dooksport.com/")
     }
+
 }